Consolidate plugins package into host package
